test: fix failing test
The test "Room BREAD >  When the last room is deleted > There should be no more players in the database" is failing when we run the whole test suite because the room/player created in the very first fixture hasn't been deleted.

By placing the fixture under a subclass, the fixture won't affect the other tests.

Closes DIC-44